update: SkillTables, SkillTableコンポーネントのデモ表示を追加
