+++ src/components/SkillTables/SkillTables.tsx
@@ -1,85 +1,13 @@
 import React from 'react'
-import { SkillTable } from '@/components'
 
-const reactSteps = [
-  'create-react-app',
-  'tsx',
-  'react-router-dom',
-  'material ui',
-  'Custom Hooks',
-  'styled-components',
-  'React 18.x Features',
-  'OAuth',
-  'Unit Test',
-]
+type Props = {
+  children?: React.ReactNode
+}
 
-const hooksSteps = ['useState', 'useEffect', 'useRef', 'useCallback', 'useContext', 'useMemo', 'useReducer']
+const SkillTables: React.FC<Props> = (props: Props) => {
+  const { children } = props
 
-const reduxSteps = ['flux-flow', 'ducks-pattern', 're-ducks-pattern', 'redux-thunk', 'redux-saga']
-
-const vueSteps = ['Atomic Design', 'Vuetify', 'Vue-Router', 'Vuex', 'Vue 3.x Features']
-
-const nodeSteps = ['Connect DB', 'CRUD', 'OAuth', 'NPM Version Management', 'GraphQL', 'Apollo']
-
-const openApiSteps = ['API Design', 'operationId Definition', 'OpenAPI Extensions']
-
-const SkillTables: React.FC = () => {
-  const frontEndProps = [
-    {
-      name: 'React.js',
-      steps: reactSteps,
-      activeStep: 5,
-    },
-    {
-      name: 'React Hooks',
-      steps: hooksSteps,
-      activeStep: 5,
-    },
-    {
-      name: 'Redux',
-      steps: reduxSteps,
-      activeStep: 3,
-    },
-    {
-      name: 'Vue.js',
-      steps: vueSteps,
-      activeStep: 4,
-    },
-  ]
-
-  const backEndProps = [
-    {
-      name: 'Node.js',
-      steps: nodeSteps,
-      activeStep: 2,
-    },
-    {
-      name: 'OpenAPI',
-      steps: openApiSteps,
-      activeStep: 1,
-    },
-  ]
-
-  return (
-    <>
-      {frontEndProps && backEndProps ? (
-        <>
-          <SkillTable
-            title="Front-End Goal Image"
-            link="https://github.com/Yuisei-Maruyama/MyPortfolio"
-            frontEndProps={frontEndProps}
-          />
-          <SkillTable
-            title="Back-End Goal Image"
-            link="https://github.com/Yuisei-Maruyama/MyPortfolio_Backend"
-            backEndProps={backEndProps}
-          />
-        </>
-      ) : (
-        <></>
-      )}
-    </>
-  )
+  return <>{children}</>
 }
 
 export default SkillTables
